+++ tests/test_pygeoprocessing_1_0.py
@@ -15,9 +15,13 @@ import scipy.ndimage
 import pygeoprocessing
 import pygeoprocessing.testing
 from pygeoprocessing.testing import sampledata
-import pygeoprocessing.routing
 import shapely.geometry
 import mock
+
+try:
+    from builtins import reload
+except ImportError:
+    from imp import reload
 
 
 class PyGeoprocessing10(unittest.TestCase):
@@ -67,8 +71,8 @@ class PyGeoprocessing10(unittest.TestCase):
         with mock.patch('pygeoprocessing.pkg_resources.get_distribution',
                         side_effect=DistributionNotFound('pygeoprocessing')):
             with self.assertRaises(RuntimeError):
-                # RuntimeError is a side effect of `import taskgraph`, so we
-                # reload it to retrigger the metadata load.
+                # RuntimeError is a side effect of `import pygeoprocessing`,
+                # so we reload it to retrigger the metadata load.
                 pygeoprocessing = reload(pygeoprocessing)
 
     def test_reclassify_raster_missing_pixel_value(self):
@@ -90,10 +94,15 @@ class PyGeoprocessing10(unittest.TestCase):
             test_value: 100,
         }
         target_nodata = -1
-        with self.assertRaises(ValueError):
+        with self.assertRaises(ValueError) as cm:
             pygeoprocessing.reclassify_raster(
                 (raster_path, 1), value_map, target_path, gdal.GDT_Float32,
                 target_nodata, values_required=True)
+        expected_message = (
+            'The following 1 raster values [-0.5] from "%s" do not have ' %
+            (raster_path,))
+        actual_message = str(cm.exception)
+        self.assertTrue(expected_message in actual_message, actual_message)
 
     def test_reclassify_raster(self):
         """PGP.geoprocessing: test reclassify raster."""
@@ -678,7 +687,7 @@ class PyGeoprocessing10(unittest.TestCase):
             base_a_raster_info['pixel_size'], bounding_box_mode,
             base_vector_path_list=None, raster_align_index=0)
 
-        for raster_index in xrange(2):
+        for raster_index in range(2):
             target_raster_info = pygeoprocessing.get_raster_info(
                 target_raster_path_list[raster_index])
             target_raster = gdal.Open(target_raster_path_list[raster_index])
@@ -736,7 +745,7 @@ class PyGeoprocessing10(unittest.TestCase):
             raster_align_index=0, base_vector_path_list=[single_pixel_path])
 
         expected_matrix = numpy.ones((1, 1), numpy.int16)
-        for raster_index in xrange(2):
+        for raster_index in range(2):
             target_raster_info = pygeoprocessing.get_raster_info(
                 target_raster_path_list[raster_index])
             target_raster = gdal.Open(target_raster_path_list[raster_index])
@@ -901,7 +910,7 @@ class PyGeoprocessing10(unittest.TestCase):
             nodata_target, reference.pixel_size(30), filename=base_path)
 
         target_path = os.path.join(
-            self.workspace_dir, 'target.tif')
+            self.workspace_dir, 'subdir', 'target.tif')
         pygeoprocessing.raster_calculator(
             [(base_path, 1)], lambda x: x, target_path,
             gdal.GDT_Int32, nodata_target, calc_raster_stats=True)
@@ -922,10 +931,15 @@ class PyGeoprocessing10(unittest.TestCase):
         for bad_raster_path_band_list in [
                 [base_path], [(base_path, "1")], [(1, 1)],
                 [(base_path, 1, base_path, 2)], base_path]:
-            with self.assertRaises(ValueError):
+            with self.assertRaises(ValueError) as cm:
                 pygeoprocessing.raster_calculator(
                     bad_raster_path_band_list, lambda x: x, target_path,
                     gdal.GDT_Int32, nodata_target, calc_raster_stats=True)
+            expected_message = (
+                'Expected a list of path / integer band tuples, numbers, or')
+            actual_message = str(cm.exception)
+            self.assertTrue(
+                expected_message in actual_message, actual_message)
 
     def test_raster_calculator_no_path(self):
         """PGP.geoprocessing: raster_calculator raise ex. on bad file path."""
@@ -933,10 +947,14 @@ class PyGeoprocessing10(unittest.TestCase):
         nonexistant_path = os.path.join(self.workspace_dir, 'nofile.tif')
         target_path = os.path.join(
             self.workspace_dir, 'target.tif')
-        with self.assertRaises(ValueError):
+        with self.assertRaises(ValueError) as cm:
             pygeoprocessing.raster_calculator(
                 [(nonexistant_path, 1)], lambda x: x, target_path,
                 gdal.GDT_Int32, nodata_target, calc_raster_stats=True)
+        expected_message = (
+            "The following files were expected but do not exist on the ")
+        actual_message = str(cm.exception)
+        self.assertTrue(expected_message in actual_message, actual_message)
 
     def test_raster_calculator_nodata(self):
         """PGP.geoprocessing: raster_calculator test with all nodata."""
@@ -966,11 +984,14 @@ class PyGeoprocessing10(unittest.TestCase):
             [pixel_matrix], reference.origin, reference.projection,
             nodata_base, reference.pixel_size(30), filename=base_path)
 
-        with self.assertRaises(ValueError):
+        with self.assertRaises(ValueError) as cm:
             # intentionally passing target path as base path to raise error
             pygeoprocessing.raster_calculator(
                 [(base_path, 1)], lambda x: x, base_path,
                 gdal.GDT_Int32, nodata_base, calc_raster_stats=True)
+        expected_message = 'is used as a target path, but it is also in the '
+        actual_message = str(cm.exception)
+        self.assertTrue(expected_message in actual_message)
 
     def test_rs_calculator_bad_overlap(self):
         """PGP.geoprocessing: rs_calculator expected error on bad overlap."""
@@ -989,13 +1010,235 @@ class PyGeoprocessing10(unittest.TestCase):
             nodata_base, reference.pixel_size(30), filename=base_path_b)
 
         target_path = os.path.join(self.workspace_dir, 'target.tif')
-        with self.assertRaises(ValueError):
-            # intentionally passing a filename rather than a list of files
-            # to get an expected exception
+        with self.assertRaises(ValueError) as cm:
             pygeoprocessing.raster_calculator(
                 [(base_path_a, 1), (base_path_b, 1)], lambda x: x,
                 target_path, gdal.GDT_Int32, nodata_base,
                 gtiff_creation_options=None, calc_raster_stats=True)
+        expected_message = 'Input Rasters are not the same dimensions.'
+        actual_message = str(cm.exception)
+        self.assertTrue(expected_message in actual_message)
+
+    def test_raster_calculator_constant_args_error(self):
+        """PGP.geoprocessing: handle empty input arrays."""
+        import pygeoprocessing
+
+        target_path = os.path.join(self.workspace_dir, 'target.tif')
+        with self.assertRaises(ValueError) as cm:
+            # no input args should cause a ValueError
+            pygeoprocessing.raster_calculator(
+                [], lambda: None, target_path,
+                gdal.GDT_Float32, None)
+        expected_message = '`base_raster_path_band_const_list` is empty'
+        actual_message = str(cm.exception)
+        self.assertTrue(expected_message in actual_message)
+
+    def test_raster_calculator_invalid_numpy_array(self):
+        """PGP.geoprocessing: handle invalid numpy array sizes."""
+        import pygeoprocessing
+
+        target_path = os.path.join(self.workspace_dir, 'target.tif')
+
+        with self.assertRaises(ValueError) as cm:
+            pygeoprocessing.raster_calculator(
+                [numpy.empty((3, 3, 3))], lambda x: None, target_path,
+                gdal.GDT_Float32, None)
+        expected_message = 'Numpy array inputs must be 2 dimensions or less'
+        actual_message = str(cm.exception)
+        self.assertTrue(expected_message in actual_message)
+
+    def test_raster_calculator_invalid_band_numbers(self):
+        """PGP.geoprocessing: ensure invalid band numbers fail."""
+        import pygeoprocessing
+
+        driver = gdal.GetDriverByName("GTiff")
+        base_path = os.path.join(self.workspace_dir, 'base.tif')
+        new_raster = driver.Create(
+            base_path, 128, 128, 1, gdal.GDT_Int32,
+            options=(
+                'TILED=YES', 'BLOCKXSIZE=16', 'BLOCKYSIZE=16'))
+        new_raster.FlushCache()
+        new_raster = None
+
+        target_path = os.path.join(self.workspace_dir, 'target.tif')
+        with self.assertRaises(ValueError) as cm:
+            # no input args should cause a ValueError
+            pygeoprocessing.raster_calculator(
+                [(base_path, 2)], lambda: None, target_path,
+                gdal.GDT_Float32, None)
+        expected_message = "do not contain requested band "
+        actual_message = str(cm.exception)
+        self.assertTrue(expected_message in actual_message)
+
+        with self.assertRaises(ValueError) as cm:
+            # no input args should cause a ValueError
+            pygeoprocessing.raster_calculator(
+                [(base_path, 0)], lambda: None, target_path,
+                gdal.GDT_Float32, None)
+        expected_message = "do not contain requested band "
+        actual_message = str(cm.exception)
+        self.assertTrue(expected_message in actual_message)
+
+    def test_raster_calculator_unbroadcastable_array(self):
+        """PGP.geoprocessing: incompatable array sizes raise error."""
+        import pygeoprocessing
+
+        target_path = os.path.join(self.workspace_dir, 'target.tif')
+        a_arg = 3
+        x_arg = numpy.array(range(2))
+        y_arg = numpy.array(range(3)).reshape((3, 1))
+        z_arg = numpy.ones((4, 4))
+        with self.assertRaises(ValueError) as cm:
+            pygeoprocessing.raster_calculator(
+                [a_arg, x_arg, y_arg, z_arg], lambda a, x, y, z: a*x*y*z,
+                target_path, gdal.GDT_Float32, None)
+        expected_message = "inputs cannot be broadcast into a single shape"
+        actual_message = str(cm.exception)
+        self.assertTrue(expected_message in actual_message)
+
+    def test_raster_calculator_array_raster_mismatch(self):
+        """PGP.geoprocessing: bad array shape with raster raise error."""
+        target_path = os.path.join(self.workspace_dir, 'target.tif')
+        driver = gdal.GetDriverByName('GTiff')
+        base_path = os.path.join(self.workspace_dir, 'base.tif')
+        new_raster = driver.Create(
+            base_path, 128, 128, 1, gdal.GDT_Int32,
+            options=(
+                'TILED=YES', 'BLOCKXSIZE=16', 'BLOCKYSIZE=16'))
+        new_raster.GetRasterBand(1).WriteArray(
+            numpy.ones((128, 128)))
+        new_raster.FlushCache()
+        new_raster = None
+
+        z_arg = numpy.ones((4, 4))
+        with self.assertRaises(ValueError) as cm:
+            pygeoprocessing.raster_calculator(
+                [(base_path, 1), z_arg], lambda a, z: a*z,
+                target_path, gdal.GDT_Float32, None)
+        expected_message = (
+            'Raster size (128, 128) cannot be broadcast '
+            'to numpy shape (4, 4)')
+        actual_message = str(cm.exception)
+        self.assertTrue(expected_message in actual_message, actual_message)
+
+        y_arg = numpy.ones((4,))
+        with self.assertRaises(ValueError) as cm:
+            pygeoprocessing.raster_calculator(
+                [(base_path, 1), y_arg], lambda a, y: a*y,
+                target_path, gdal.GDT_Float32, None)
+        expected_message = (
+            'Raster size (128, 128) cannot be broadcast '
+            'to numpy shape (4,)')
+        actual_message = str(cm.exception)
+        self.assertTrue(expected_message in actual_message, actual_message)
+
+    def test_raster_calculator_constant_args(self):
+        """PGP.geoprocessing: test constant arguments of raster calc."""
+        import pygeoprocessing
+
+        target_path = os.path.join(self.workspace_dir, 'target.tif')
+        a_arg = 3
+        x_arg = numpy.array(range(2))
+        y_arg = numpy.array(range(3)).reshape((3, 1))
+        z_arg = numpy.ones((3, 2))
+        pygeoprocessing.raster_calculator(
+            [a_arg, x_arg, y_arg, z_arg], lambda a, x, y, z: a*x*y*z,
+            target_path, gdal.GDT_Float32, 0)
+
+        target_raster = gdal.OpenEx(target_path, gdal.OF_RASTER)
+        target_array = target_raster.GetRasterBand(1).ReadAsArray()
+        target_raster = None
+        expected_result = numpy.array([[0, 0], [0, 3], [0, 6]])
+        numpy.testing.assert_array_almost_equal(target_array, expected_result)
+
+        target_path = os.path.join(self.workspace_dir, 'target_a.tif')
+        with self.assertRaises(ValueError) as cm:
+            # this will return a scalar, when it should return 2d array
+            pygeoprocessing.raster_calculator(
+                [a_arg], lambda a: a, target_path,
+                gdal.GDT_Float32, None)
+        expected_message = (
+            "Expected `local_op` to return a numpy.ndarray")
+        actual_message = str(cm.exception)
+        self.assertTrue(expected_message in actual_message)
+
+        # try the scalar as a 2d array
+        pygeoprocessing.raster_calculator(
+            [a_arg], lambda a: numpy.array([[a]]), target_path,
+            gdal.GDT_Float32, None)
+
+        target_raster = gdal.OpenEx(target_path, gdal.OF_RASTER)
+        target_array = target_raster.GetRasterBand(1).ReadAsArray()
+        target_raster = None
+        expected_result = numpy.array([[a_arg]])
+        numpy.testing.assert_array_almost_equal(target_array, expected_result)
+
+        target_path = os.path.join(
+            self.workspace_dir, 'target_1d_2darray.tif')
+        pygeoprocessing.raster_calculator(
+            [y_arg], lambda y: y, target_path,
+            gdal.GDT_Float32, None)
+
+        target_raster = gdal.OpenEx(target_path, gdal.OF_RASTER)
+        target_array = target_raster.GetRasterBand(1).ReadAsArray()
+        target_raster = None
+        numpy.testing.assert_array_almost_equal(target_array, y_arg)
+
+        target_path = os.path.join(self.workspace_dir, 'target_1d_only.tif')
+        pygeoprocessing.raster_calculator(
+            [x_arg], lambda x: x, target_path,
+            gdal.GDT_Float32, None)
+
+        target_raster = gdal.OpenEx(target_path, gdal.OF_RASTER)
+        target_array = target_raster.GetRasterBand(1).ReadAsArray()
+        target_raster = None
+        numpy.testing.assert_array_almost_equal(
+            target_array, x_arg.reshape((1, x_arg.size)))
+
+    def test_combined_constant_args_raster(self):
+        """PGP.geoprocessing: test raster calc with constant args."""
+        driver = gdal.GetDriverByName('GTiff')
+        base_path = os.path.join(self.workspace_dir, 'base.tif')
+        new_raster = driver.Create(
+            base_path, 128, 128, 1, gdal.GDT_Int32,
+            options=(
+                'TILED=YES', 'BLOCKXSIZE=32', 'BLOCKYSIZE=32'))
+        new_band = new_raster.GetRasterBand(1)
+        nodata = 0
+        new_band.SetNoDataValue(nodata)
+        raster_array = numpy.ones((128, 128))
+        raster_array[127, 127] = nodata
+        new_band.WriteArray(raster_array)
+        raster_array = None
+        new_band.FlushCache()
+        new_band = None
+        new_raster.FlushCache()
+        new_raster = None
+
+        target_path = os.path.join(self.workspace_dir, 'target.tif')
+
+        # making a local op that needs a valid mask to ensure that `col_array`
+        # is tiled out correctly
+        def local_op(scalar, raster_array, col_array):
+            valid_mask = raster_array != nodata
+            result = numpy.empty_like(raster_array)
+            result[:] = nodata
+            result[valid_mask] = (
+                scalar * raster_array[valid_mask] * col_array[valid_mask])
+            return result
+
+        pygeoprocessing.raster_calculator(
+            [10, (base_path, 1), numpy.array(range(128))],
+            local_op, target_path, gdal.GDT_Float32, None, largest_block=0)
+
+        target_raster = gdal.OpenEx(target_path, gdal.OF_RASTER)
+        result = target_raster.GetRasterBand(1).ReadAsArray()
+
+        expected_result = (
+            10 * numpy.ones((128, 128)) * numpy.array(range(128)))
+        # we expect one pixel to have been masked out
+        expected_result[127, 127] = nodata
+        numpy.testing.assert_allclose(result, expected_result)
 
     def test_new_raster_from_base_unsigned_byte(self):
         """PGP.geoprocessing: test that signed byte rasters copy over."""
@@ -1397,10 +1640,15 @@ class PyGeoprocessing10(unittest.TestCase):
             [kernel_array], reference.origin, reference.projection,
             nodata_target, reference.pixel_size(30), filename=kernel_path)
         target_path = os.path.join(self.workspace_dir, 'target.tif')
-        with self.assertRaises(ValueError):
+        with self.assertRaises(ValueError) as cm:
             pygeoprocessing.convolve_2d(
                 (signal_path, 1), (kernel_path, 1), target_path,
                 target_datatype=gdal.GDT_Int32)
+        expected_message = (
+            "`target_datatype` is set, but `target_nodata` is None. ")
+        actual_message = str(cm.exception)
+        self.assertTrue(expected_message in actual_message)
+
 
     def test_convolve_2d_reverse(self):
         """PGP.geoprocessing: test convolve 2d reversed."""
@@ -1449,7 +1697,7 @@ class PyGeoprocessing10(unittest.TestCase):
         kernel_path = os.path.join(self.workspace_dir, 'kernel.tif')
         kernel_array = numpy.zeros(
             (n_kernel_pixels, n_kernel_pixels), numpy.float32)
-        kernel_array[n_kernel_pixels/2,n_kernel_pixels/2] = 1
+        kernel_array[int(n_kernel_pixels/2),int(n_kernel_pixels/2)] = 1
         pygeoprocessing.testing.create_raster_on_disk(
             [kernel_array], reference.origin, reference.projection,
             nodata_target, reference.pixel_size(30), filename=kernel_path)
@@ -1475,7 +1723,7 @@ class PyGeoprocessing10(unittest.TestCase):
         pixel_matrix[:] = numpy.arange((n_pixels))
         nodata_value = -1
         # make a nodata hole in the middle to test boundary cases
-        pixel_matrix[n_pixels/2, n_pixels/2] = nodata_value
+        pixel_matrix[int(n_pixels/2), int(n_pixels/2)] = nodata_value
         dem_path = os.path.join(self.workspace_dir, 'dem.tif')
         target_slope_path = os.path.join(self.workspace_dir, 'slope.tif')
         pygeoprocessing.testing.create_raster_on_disk(
@@ -1577,10 +1825,15 @@ class PyGeoprocessing10(unittest.TestCase):
             fields={'id': 'int'}, attributes=[{'id': 5}],
             vector_format='GeoJSON', filename=base_vector_path)
 
-        with self.assertRaises(ValueError):
+        with self.assertRaises(ValueError) as cm:
             pygeoprocessing.rasterize(
                 base_vector_path, target_raster_path, [test_value], None,
                 layer_index=0)
+        expected_message = (
+            "%s doesn't exist, but needed to rasterize." % target_raster_path)
+        actual_message = str(cm.exception)
+        self.assertTrue(expected_message in actual_message, actual_message)
+
 
     def test_distance_transform_edt(self):
         """PGP.geoprocessing: test distance transform EDT."""
@@ -1588,7 +1841,7 @@ class PyGeoprocessing10(unittest.TestCase):
         n_pixels = 1000
         base_raster_array = numpy.zeros(
             (n_pixels, n_pixels), dtype=numpy.int8)
-        base_raster_array[n_pixels/2, n_pixels/2] = 1
+        base_raster_array[int(n_pixels/2), int(n_pixels/2)] = 1
         base_raster_array[0, 0] = 1
         base_raster_array[0, n_pixels-1] = 1
         base_raster_array[3, 4] = 1
@@ -1602,9 +1855,9 @@ class PyGeoprocessing10(unittest.TestCase):
         base_raster_array[5, 6] = 1
         base_raster_array[n_pixels-1, 0] = 1
         base_raster_array[n_pixels-1, n_pixels-1] = 1
-        base_raster_array[n_pixels/2, n_pixels/2] = 1
-        base_raster_array[n_pixels/2, n_pixels/4] = 1
-        base_raster_array[n_pixels/2, (3*n_pixels)/4] = 1
+        base_raster_array[int(n_pixels/2), int(n_pixels/2)] = 1
+        base_raster_array[int(n_pixels/2), int(n_pixels/4)] = 1
+        base_raster_array[int(n_pixels/2), int((3*n_pixels)/4)] = 1
         nodata_target = -1
         base_raster_path = os.path.join(
             self.workspace_dir, 'base_raster.tif')
@@ -1808,3 +2061,44 @@ class PyGeoprocessing10(unittest.TestCase):
 
         numpy.testing.assert_almost_equal(target_array, expected_array)
         self.assertEqual(nodata_value, 0)
+
+    def test_align_with_target_sr(self):
+        """PGP: test align_and_resize_raster_stack with a target sr."""
+
+        wgs84_sr = osr.SpatialReference()
+        wgs84_sr.ImportFromEPSG(4326)  # WGS84 EPSG
+
+        driver = gdal.GetDriverByName("GTiff")
+        base_path = os.path.join(self.workspace_dir, 'base.tif')
+        new_raster = driver.Create(base_path, 10, 10, 1, gdal.GDT_Int32)
+        new_raster.FlushCache()
+
+        new_raster.SetProjection(wgs84_sr.ExportToWkt())
+        new_raster.SetGeoTransform([
+            -123.587984, 1.0, 0.0, 44.725814, 0.0, -1.0])
+        new_raster = None
+
+        target_path = os.path.join(self.workspace_dir, 'target.tif')
+
+        target_ref = osr.SpatialReference()
+        target_ref.ImportFromEPSG(26910)  # UTM10N EPSG
+
+        pygeoprocessing.align_and_resize_raster_stack(
+            [base_path], [target_path], ['nearest'],
+            (3e4, -3e4), 'intersection', target_sr_wkt=target_ref.ExportToWkt())
+
+        target_raster_info = pygeoprocessing.get_raster_info(target_path)
+
+        # I have confidence in the upper left bounding box coordinate
+        # hardcoded below because I plotted the WGS84 and UTM10N rasters on
+        # top of each other and moused over the upper right hand corner.
+        # Note the warping of wgs84 to utm will cause distortion and this
+        # function attempts to make a raster that bounds that distortion
+        # without losing any data, so a direct mapping from the lat/lng
+        # coordinate to the one below will be slightly of because the
+        # warped raster's bounding box will be a little larger.
+        self.assertIs(
+            numpy.testing.assert_allclose(
+                [446166.79245811916, 4995771.240781772],
+                [target_raster_info['bounding_box'][0],
+                 target_raster_info['bounding_box'][3]]), None)